refactor 플레이 결과 저장 (step 3 피드백)
